Initial shot at assets REST API
Currently supports only simple devices (dc, rooms, rows and racks), and
reads data from static files that could be created using fake_message
tool. It is not completelly according to rfc11, some values has to be
taken to the upper layer and there is no validation of anything, but it
provides some results.

Signed-off-by: Michal Hrusecky <Michal@Hrusecky.net>