BIOS-476 BIOS-478: Implementation done; bug fixed
-> very simple "netlogic"-like component called filip
   essentialy a lackey between  netmon,... <-> simple <-> nmap
-> bug BIOS-478 fixed;
   The problem was that we didn't filter out hosts that had zero
hostname elements inside <hostnames></...>
-> in persistencelogic i initially went for an approach that transfers
ownership. Then i found out that process_message() does not support his
and i didn't have time to rewrite the others. Please keep this as it is
for the moment; will be consolidated with next PR.
-> other minor changes elsewhere in codebase
-> at the moment netmon has to perform two send operations because we
are using a dealer socket for it; I was suggesting a quick revamp to PUB
- SUB for netmon, but MHR didn't agree. If you have any problems with
  this talk to the boss ;).
-> README file with a summary of sockets and hookpoints; please try to
keep this up to date

Signed-off-by: Karol Hrdina <karolhrdina@eaton.com>