Problem: dd-mm-yyyy strings recognized as yy-mm-dd
Solution: add a support of yyyy-mm-dd and dd-mm-yyyy formats by
mandating that one of the parts must be at least 4 characters.

This change rejects dd-mm-yy and yy-mm-dd formatsm because it's
impossible to distinguish between them. The only one exception is
dd-%b-yy, where %b is locale dependent abbreciated month name and it's a
date used by Excell.

Signed-off-by: Michal Vyskocil <michalvyskocil@eaton.com>