Problem: utf8 strings are not properly exported
Solution: remove the code wrapping non printable character from
utils::json::escape. utf-8 is a default encoding for it and most of
parsers are capable to work with utf-8 just well.

https://www.ietf.org/rfc/rfc4627.txt

"""
3.  Encoding

   JSON text SHALL be encoded in Unicode.  The default encoding is
      UTF-8.
"""

This contradict escaping rules, but ...

http://stackoverflow.com/questions/583562/json-character-encoding-is-utf-8-well-supported-by-browsers-or-should-i-use-nu

"""
All conforming JSON parsers will be able to handle proper UTF-8 just as
well as the safe numeric encodings.
"""

We do not send output in anything else than utf-8, so it's fine.

Signed-off-by: Michal Vyskocil <michalvyskocil@eaton.com>