Problem: have no idea where to put stripping function
Solution: drop (most of) the mess with various libpriv, particularly
merge utils and db. This will a) simplify the Makefile and add a linker
a chance to do LTO or --as-needed optimalizations. In current forms all
libprivs are used everywhere anyway, so there is little point to keep
the separated.