Problem: use of persist::name_to_asset_id() is inconsistent
Context: the recently introduced persist::name_to_asset_id() returns an int64_t value which is either "-1" for errors (and/or absent asset), or the uint32_t value of database internal identifier. Some consumers just casted the return value to uint32_t and/or did not check for "-1" returns, and got funny results when the value was actually "-1" for error.

Solution: revise existing calls to the function to use int64_t, check for "-1", and use the rest as before. Also make the code look more consistent by naming the value "dbid" in more places (where it causes just little changes in codebase).